+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog (niondata)
 UNRELEASED
 ----------
 
+- Add a get_component method to Registry for easier lookup.
+
 - Treat '.' in float numbers as decimal point independent of locale when parsing, leave locale decimal point valid too.
 
 0.3.10 (2018-05-10)